fix(obsidian): add tests for and address #1277
