--- packages/obsidian-plugin/src/index.test.ts
@@ -1,5 +1,0 @@
-import { expect, test } from 'vitest';
-
-test('adding', () => {
-	expect(1 + 1).toBe(2);
-});
